rename(rpc): rpc.rs to methods.rs
