+++ zebra-rpc/src/lib.rs
@@ -4,4 +4,4 @@
 #![doc(html_logo_url = "https://www.zfnd.org/images/zebra-icon.png")]
 #![doc(html_root_url = "https://doc.zebra.zfnd.org/zebra_rpc")]
 
-pub mod rpc;
+pub mod methods;
